requirements/dev.txt
# Include packages for all environments
-r base.txt

# Only include these packages in dev
djangorestframework
markdown
django-filter
